Remove redundant options field
This adds a empty base class for options that both NounOptions and
NumericOptions subclass, so as to allow the directive to contain a
single field of Options, that is cast to the correct subtype when
needed.

Attempting to set the wrong kind of options for a directive will cause
an exception to be thrown.